Rozbicie odpowiedzialnosci klasy paginacji gorliwej
- dodanie nowej klasy filtrującej paginacje gorliwą
